rdf_owl/requirements.txt
rdflib==7.0.0
pydantic==2.5.0
requests==2.31.0
PyYAML==6.0.1
